Move wcs_from_pointing to a function.
