+++ gwcs/__init__.py
@@ -17,5 +17,6 @@ from ._astropy_init import *
 # For egg_info test builds to pass, put package imports here.
 if not _ASTROPY_SETUP_:
     from .wcs import *
+    from .wcstools import *
     from .coordinate_frames import *
     from .selector import *
